feat: added websocket ✨

+++ src/controllers/feed.ts
@@ -5,6 +5,8 @@ import { validationResult } from 'express-validator'
 import { NextFunction, Request, Response } from 'express'
 import { CustomError } from '../types/utils'
 import Post from '../models/post'
+import User from '../models/user'
+import { getIo } from '../libs/socket'
 
 type Creator = {
     name: string
@@ -37,6 +39,8 @@ export const getPosts = async (
 
     totalItems = count
     const posts = await Post.find()
+        .populate('creator')
+        .sort({ createdAt: -1 })
         .skip((currentPage - 1) * PER_PAGE)
         .limit(PER_PAGE)
         .catch((err: CustomError) => {
@@ -74,11 +78,7 @@ export const getPost = async (
     }
 }
 
-export const createPost = async (
-    req: Request,
-    res: Response,
-    next: NextFunction
-): Promise<void> => {
+export const createPost = async (req, res, next): Promise<void> => {
     const errors = validationResult(req)
 
     if (!errors.isEmpty()) {
@@ -101,7 +101,7 @@ export const createPost = async (
         title: req.body.title,
         content: req.body.content,
         imageURL,
-        creator: { name: 'Test User' }
+        creator: req.userId
     })
 
     const result = await post.save().catch((err: CustomError) => {
@@ -111,9 +111,28 @@ export const createPost = async (
         next(err)
     })
 
+    const user = await User.findById(req.userId).catch(err => {
+        console.error(err)
+        next(err)
+    })
+
+    getIo().emit('posts', {
+        action: 'create',
+        post: { ...post._doc, creator: { _id: req.userId, name: user.name } }
+    })
+
+    user.posts.push(post)
+    await user.save().catch((err: CustomError) => {
+        if (!err.statusCode) {
+            err.statusCode = 500
+        }
+        next(err)
+    })
+
     res.status(201).json({
         message: 'success!',
-        post: result
+        creator: { _id: user._id, name: user.name },
+        post
     })
 }
 
@@ -141,25 +160,37 @@ export const updatePost = async (req, res, next): Promise<void> => {
         next(err)
     }
 
-    const post = await Post.findById(postId).catch(err => {
-        console.error(err)
-        next(err)
-    })
+    const post = await Post.findById(postId)
+        .populate('creator')
+        .catch(err => {
+            console.error(err)
+            next(err)
+        })
 
     if (!post) {
         const err: CustomError = new Error('Could not find a post.')
         err.statusCode = 404
         next(err)
     } else {
-        if (imageURL !== post.imageURL) {
-            clearImage(post.imageURL)
-        }
+        if (post.creator._id.toString() !== req.userId.toString()) {
+            const err: CustomError = new Error('Not authorized.')
+            err.statusCode = 403
+            next(err)
+            throw err
+        } else {
+            if (imageURL !== post.imageURL) {
+                clearImage(post.imageURL)
+            }
 
-        post.title = title
-        post.imageURL = imageURL
-        post.content = content
-        post.save()
-        res.status(200).json({ post })
+            post.title = title
+            post.imageURL = imageURL
+            post.content = content
+            const result = post.save()
+
+            getIo().emit('posts', { action: 'update', post: result })
+
+            res.status(200).json({ post })
+        }
     }
 }
 
@@ -175,19 +206,41 @@ export const deletePost = async (req, res, next): Promise<void> => {
         err.statusCode = 404
         next(err)
     } else {
-        //ToDo: Check logged in user
-
-        const result = await Post.findByIdAndDelete(postId).catch(err => {
-            console.error(err)
+        if (post.creator.toString() !== req.userId.toString()) {
+            const err: CustomError = new Error('Not authorized.')
+            err.statusCode = 403
             next(err)
-        })
-        if (result) {
-            clearImage(post.imageURL)
-            console.info(result)
-            res.status(200).json({
-                message: 'deleted post.',
-                id: postId
+            throw err
+        } else {
+            const result = await Post.findByIdAndDelete(postId).catch(err => {
+                console.error(err)
+                next(err)
             })
+            if (result) {
+                const user = await User.findById(req.userId).catch(err => {
+                    console.error(err)
+                    next(err)
+                })
+
+                if (user) {
+                    user.posts.pull(postId)
+                    await user.save().catch(err => {
+                        console.error(err)
+                        next(err)
+                    })
+                    clearImage(post.imageURL)
+                    console.info(result)
+                    getIo().emit('posts', { action: 'delete', post: postId })
+                    res.status(200).json({
+                        message: 'deleted post.',
+                        id: postId
+                    })
+                } else {
+                    const error: CustomError = new Error('Not authorized')
+                    error.statusCode = 403
+                    next(error)
+                }
+            }
         }
     }
 }
